pi/requirements.txt
python-dotenv
requests
sounddevice
numpy
